Ребрендинг в отражателей проджектайлов. Ище чут чут улучшил код
